Ограничен доступ в удалении и добавлении.

+++ pis_c/Controllers/ModelController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using pis_c.Models.DBEntities;
@@ -9,6 +11,7 @@ using pis_c.ViewModels;
 
 namespace pis_c.Controllers
 {
+    [Authorize]
     public class ModelController : Controller
     {
         AppDbContext dbContext;
@@ -20,12 +23,16 @@ namespace pis_c.Controllers
 
         public IActionResult All()
         {
+            if (!IsAdmin(HttpContext.User))
+                return RedirectToAction("Index", "Home");
             return View(dbContext.Models.Include(m => m.Brand).OrderBy(m => m.Brand.Name));
         }
 
         [HttpGet]
         public IActionResult Add()
         {
+            if (!IsAdmin(HttpContext.User))
+                return RedirectToAction("Index", "Home");
             ViewBag.Brands = dbContext.Brands.ToList();
             return View();
         }
@@ -33,6 +40,8 @@ namespace pis_c.Controllers
         [HttpPost]
         public IActionResult Add(ModelViewModel model)
         {
+            if (!IsAdmin(HttpContext.User))
+                return RedirectToAction("Index", "Home");
             if (ModelState.IsValid)
             {
                 dbContext.Models.Add(new Model()
@@ -48,15 +57,24 @@ namespace pis_c.Controllers
 
         public IActionResult Delete(int modelId)
         {
+            if (!IsAdmin(HttpContext.User))
+                return RedirectToAction("Index", "Home");
             dbContext.Models.Remove(dbContext.Models.FirstOrDefault(m => m.Id == modelId));
             dbContext.SaveChanges();
             return RedirectToAction("All");
         }
 
+        [AllowAnonymous]
         [HttpGet]
         public IActionResult ModelsHtmlOptions(int brandId)
         {
             return PartialView(dbContext.Models.Where(m => m.BrandId == brandId).ToList());
+        }
+
+        private bool IsAdmin(ClaimsPrincipal user)
+        {
+            var email = user.Claims.ElementAt(0).Value;
+            return dbContext.Users.FirstOrDefault(u => u.Email == email).IsAdmin;
         }
     }
 }
